@@ -2,7 +2,7 @@
   "name": "react-native-ui-lib",
   "version": "4.2.0",
   "main": "index.js",
-  "typings": "typings/index.d.ts",
+  "types": "index.d.ts",
   "author": "Ethan Sharabi <ethan.shar@gmail.com>",
   "homepage": "https://github.com/wix/react-native-ui-lib",
   "license": "MIT",
@@ -15,6 +15,7 @@
   },
   "scripts": {
     "start": "watchman watch-del-all && export DEV_MODE=true && react-native start",
+    "start-ios": "node ./scripts/cliStuff.js",
     "ios": "react-native run-ios",
     "android": "react-native run-android",
     "pretest": "npm run lint",
@@ -24,7 +25,7 @@
     "lint:fix": "eslint src -c .eslintrc.js --fix",
     "lint:test": "mocha --compilers js:babel-core/register eslint-rules/tests/lib/rules",
     "xcode": "open ./ios/uilib.xcodeproj",
-    "build": "mkdir -p dist && cp -r src/assets dist/ && babel src --out-dir dist",
+    "build": "tsc",
     "log": "react-native log-ios | grep 'ethan -'",
     "docs:install": "(cd ./uilib-docs && rm -rf node_modules && rm -rf package-lock.json && npm install)",
     "docs:deploy": "(cd ./uilib-docs && gatsby build --prefix-paths && gh-pages -d public --branch gh-pages)",
@@ -60,9 +61,14 @@
     "@react-native-community/blur": "^3.4.1",
     "@react-native-community/datetimepicker": "^2.1.0",
     "@react-native-community/eslint-config": "latest",
+    "@react-native-community/netinfo": "^5.6.2",
+    "@types/lodash": "^4.0.0",
     "@types/prop-types": "^15.5.3",
-    "@types/react-native": "^0.55.21",
+    "@types/react-native": ">=0.61.5",
+    "@types/tinycolor2": "^1.4.2",
     "@welldone-software/why-did-you-render": "^3.2.1",
+    "@typescript-eslint/eslint-plugin": "^2.13.0",
+    "@typescript-eslint/parser": "^2.13.0",
     "babel-jest": "^24.5.0",
     "babel-preset-react-native": "*",
     "eslint": "^5.16.0",
@@ -84,15 +90,16 @@
     "react-native-navigation": "^3.4.0",
     "react-native-reanimated": "^1.7.0",
     "shell-utils": "^1.0.10",
-    "typescript": "^2.9.2"
+    "typescript": "^3.8.3"
   },
   "peerDependencies": {
     "react": ">=16.0.0",
-    "react-native": ">=0.51.0",
+    "react-native": ">=0.61.5",
     "react-native-gesture-handler": ">=1.5.0",
     "react-native-reanimated": ">=1.4.0",
     "@react-native-community/blur": ">=3.4.1",
-    "@react-native-community/datetimepicker": "^2.1.0"
+    "@react-native-community/datetimepicker": "^2.1.0",
+    "@react-native-community/netinfo": "^5.6.2"
   },
   "jest": {
     "preset": "react-native",
