@@ -2,38 +2,39 @@ import _ from 'lodash';
 import React, {Component} from 'react';
 import AsyncStorage from '@react-native-community/async-storage';
 import PropTypes from 'prop-types';
-import autobind from 'react-autobind';
 import {StyleSheet, FlatList, ViewPropTypes} from 'react-native';
 import {Navigation} from 'react-native-navigation';
-import {
-  ThemeManager,
-  Constants,
-  Assets,
-  Colors,
-  View,
-  Text,
-  Button,
-  Carousel,
-  TextField,
-  Image
-} from 'react-native-ui-lib'; //eslint-disable-line
+import {gestureHandlerRootHOC} from 'react-native-gesture-handler';
+import {Assets, Colors, View, Text, TouchableOpacity, TextField, Image, TabController} from 'react-native-ui-lib'; //eslint-disable-line
 import {navigationData} from './MenuStructure';
-
 
 const settingsIcon = require('../assets/icons/settings.png');
 const chevronIcon = require('../assets/icons/chevronRight.png');
 
-export default class MainScreen extends Component {
+class MainScreen extends Component {
   static propTypes = {
     containerStyle: ViewPropTypes.style,
     renderItem: PropTypes.func,
-    renderSectionTitle: PropTypes.func,
     pageStyle: ViewPropTypes.style
   };
 
+  static options() {
+    return {
+      topBar: {
+        rightButtons: [
+          {
+            id: 'uilib.settingsButton',
+            testID: 'uilib.settingsButton',
+            enabled: true,
+            icon: settingsIcon
+          }
+        ]
+      }
+    };
+  }
+
   constructor(props) {
     super(props);
-    autobind(this);
 
     const data = props.navigationData || navigationData;
 
@@ -45,70 +46,30 @@ export default class MainScreen extends Component {
     this.filterExplorerScreens = _.throttle(this.filterExplorerScreens, 300);
 
     Navigation.events().bindComponent(this);
-
-    const navigationStyle = this.getSearchNavigationStyle();
-    navigationStyle.topBar.rightButtons.push({
-      id: 'uilib.settingsButton',
-      testID: 'uilib.settingsButton',
-      enabled: true,
-      icon: settingsIcon
-    });
-    Navigation.mergeOptions(props.componentId, navigationStyle);
   }
 
-  /** Events */
-  onChangePage(newPage) {
-    this.setState({currentPage: newPage});
-  }
-
-  onSearchBoxBlur() {
+  onSearchBoxBlur = () => {
     this.closeSearchBox();
     this.filterExplorerScreens('');
-  }
+  };
 
-  /** Navigation */
-  getMenuData() {
+  getMenuData = () => {
     return this.props.navigationData || navigationData;
-  }
-
-  getSearchNavigationStyle() {
-    return {
-      topBar: {
-        drawBehind: true,
-        translucent: true,
-        rightButtons: [
-          {
-            id: 'uilib.searchButton',
-            testID: 'uilib.searchButton',
-            enabled: true,
-            icon: Assets.icons.search
-          }
-        ]
-      }
-    };
-  }
+  };
 
   navigationButtonPressed = event => {
     const {buttonId} = event;
     const data = this.getMenuData();
 
-    switch (buttonId) {
-      case 'uilib.settingsButton':
-        this.pushScreen({
-          name: 'unicorn.Settings',
-          passProps: {navigationData: data, playground: this.props.playground}
-        });
-        break;
-      case 'uilib.searchButton':
-        this.input.focus();
-        this.toggleTopBar(false);
-        break;
-      default:
-        break;
+    if (buttonId === 'uilib.settingsButton') {
+      this.pushScreen({
+        name: 'unicorn.Settings',
+        passProps: {navigationData: data, playground: this.props.playground}
+      });
     }
   };
 
-  pushScreen(options) {
+  pushScreen = options => {
     Navigation.push(this.props.componentId, {
       component: {
         name: options.name || options.screen,
@@ -122,59 +83,26 @@ export default class MainScreen extends Component {
         }
       }
     });
-  }
-
-  showScreen(options) {
-    Navigation.showModal({
-      stack: {
-        children: [
-          {
-            component: {
-              name: options.name || options.screen,
-              passProps: options.passProps,
-              options: {
-                topBar: {
-                  title: {
-                    text: options.title
-                  }
-                }
-              }
-            }
-          }
-        ]
-      }
-    });
-  }
-
-  /** Actions */
-  toggleTopBar = shouldShow => {
-    Navigation.mergeOptions(this.props.componentId, {
-      topBar: {
-        visible: shouldShow,
-        animate: true
-      }
-    });
   };
 
-  closeSearchBox() {
-    this.toggleTopBar(true);
+  closeSearchBox = () => {
     this.input.blur();
-  }
+  };
 
-  setDefaultScreen = (item) => {
+  setDefaultScreen = item => {
     AsyncStorage.setItem('uilib.defaultScreen', item.screen);
-  }
+  };
 
-  openScreen(row) {
+  openScreen = row => {
     this.closeSearchBox();
 
     setTimeout(() => {
       this.filterExplorerScreens('');
       this.pushScreen(row);
     }, 0);
-  }
+  };
 
-  filterExplorerScreens(filterText) {
+  filterExplorerScreens = filterText => {
     let filteredNavigationData = {};
     const data = this.getMenuData();
 
@@ -201,150 +129,66 @@ export default class MainScreen extends Component {
       filterText,
       filteredNavigationData
     });
-  }
+  };
 
   /** Renders */
-  renderHeader() {
+  renderSearch = () => {
     return (
-      <View row spread style={{height: Constants.isIOS ? (Constants.isIphoneX ? 80 : 60) : 56}}>
-        <TextField
-          ref={r => (this.input = r)}
-          value={this.state.filterText}
-          placeholder="Search for your component..."
-          onChangeText={this.filterExplorerScreens}
-          onBlur={this.onSearchBoxBlur}
-          style={{
-            marginTop: Constants.isIOS ? Constants.statusBarHeight + 10 : 14,
-            marginLeft: 16,
-            color: ThemeManager.primaryColor,
-            width: Constants.screenWidth - 80
-          }}
-          hideUnderline
-          floatingPlaceholder={false}
-        />
-        <Button
-          testID={'SearchButton'}
-          style={{marginRight: 16, marginTop: Constants.isIOS ? Constants.statusBarHeight : 0}}
-          iconSource={Assets.icons.search}
-          size={'small'}
-          backgroundColor={'transparent'}
-          onPress={this.onSearchBoxBlur}
-        />
-      </View>
+      <TextField
+        ref={r => (this.input = r)}
+        value={this.state.filterText}
+        placeholder="Search for your component..."
+        onChangeText={this.filterExplorerScreens}
+        onBlur={this.onSearchBoxBlur}
+        containerStyle={{padding: 16, paddingBottom: 0}}
+        style={{
+          padding: 12,
+          backgroundColor: Colors.dark80,
+          borderRadius: 8
+        }}
+        enableErrors={false}
+        hideUnderline
+        floatingPlaceholder={false}
+        text70
+        rightButtonProps={{iconSource: Assets.icons.search, style: {marginRight: 12}}}
+      />
     );
-  }
+  };
 
-  renderItem({item}) {
+  renderItem = ({item}) => {
     const {renderItem} = this.props;
 
     if (renderItem) {
       return renderItem({item}, this.openScreen);
     }
 
-    return (
-      <View centerV row paddingL-20 marginB-10>
-        <Image source={chevronIcon} style={{tintColor: Colors.dark10}} supportRTL/>
-        <Text
-          style={[item.deprecate && styles.entryTextDeprecated]}
-          dark10
-          marginL-10
-          text50
+    if (item.screen) {
+      return (
+        <TouchableOpacity
+          centerV
+          row
+          spread
+          paddingH-s5
+          paddingV-s2
           onPress={() => this.openScreen(item)}
           onLongPress={() => this.setDefaultScreen(item)}
+          activeBackgroundColor={Colors.blue40}
+          activeOpacity={1}
         >
-          {item.title}
-        </Text>
-      </View>
-    );
-  }
-
-  renderBreadcrumbs() {
-    const {currentPage} = this.state;
-    const data = this.getMenuData();
-    const sections = Object.keys(data);
-
-    return (
-      <View style={styles.breadcrumbs} row>
-        {_.map(data, (section, key) => {
-          const index = sections.indexOf(key);
-          const isLast = index === sections.length - 1;
-          return (
-            <View key={key} row centerV>
-              <Button
-                // link
-                size={Button.sizes.xSmall}
-                marginB-5
-                marginR-5={!isLast}
-                text80
-                // dark50={currentPage !== index}
-                // dark10={currentPage === index}
-                // text50
-                outline={currentPage !== index}
-                label={section.title}
-                style={{height: 30}}
-                onPress={() => this.carousel.goToPage(index)}
-              />
-              {/* {!isLast && <Text marginH-5>&middot;</Text>} */}
-            </View>
-          );
-        })}
-      </View>
-    );
-  }
-
-  renderSectionTitle(title) {
-    const {renderSectionTitle} = this.props;
-
-    if (renderSectionTitle) {
-      return renderSectionTitle(title);
+          <Text style={[item.deprecate && styles.entryTextDeprecated]} dark10 text50>
+            {item.title}
+          </Text>
+          <Image source={chevronIcon} style={{tintColor: Colors.dark10}} supportRTL/>
+        </TouchableOpacity>
+      );
+    } else {
+      return (
+        <View paddingH-s5 marginV-s1 height={20} bg-grey80>
+          <Text text80M>{item.title}</Text>
+        </View>
+      );
     }
-
-    return (
-      <View style={styles.pageTitleContainer}>
-        <Text text40 style={{alignSelf: 'flex-start'}}>
-          {title}
-        </Text>
-      </View>
-    );
-  }
-
-  renderCarousel(data) {
-    const {pageStyle} = this.props;
-    const dividerTransforms = [-10, -55, -20];
-    const dividerWidths = ['60%', '75%', '90%'];
-    const keys = _.keys(data);
-
-    return (
-      <Carousel
-        ref={carousel => (this.carousel = carousel)}
-        containerStyle={{flex: 1}}
-        onChangePage={this.onChangePage} 
-      >
-        {_.map(data, (section, key) => {
-          return (
-            <View key={key} style={[styles.page, pageStyle]}>
-              {this.renderSectionTitle(section.title)}
-              <View
-                style={[
-                  styles.pageTitleExtraDivider,
-                  {width: dividerWidths[_.indexOf(keys, key) % dividerWidths.length]},
-                  {transform: [{translateX: dividerTransforms[_.indexOf(keys, key) % dividerTransforms.length]}]}
-                ]}
-              />
-              <View flex>
-                <FlatList
-                  showsVerticalScrollIndicator={false}
-                  data={section.screens}
-                  keyExtractor={item => item.screen ? item.title : `header_${item.title}`}
-                  renderItem={this.renderItem}
-                />
-              </View>
-            </View>
-          );
-        })}
-      </Carousel>
-    );
-  }
+  };
 
   renderSearchResults(data) {
     const flatData = _.flatMap(data);
@@ -353,9 +197,33 @@ export default class MainScreen extends Component {
       <FlatList
         keyboardShouldPersistTaps="always"
         data={flatData}
+        contentContainerStyle={{paddingTop: 20}}
         keyExtractor={(item, index) => index.toString()}
         renderItem={this.renderItem}
       />
+    );
+  }
+
+  renderPages(data) {
+    const {pageStyle} = this.props;
+    let index = 0;
+    return (
+      <TabController.PageCarousel>
+        {_.map(data, (section, key) => {
+          return (
+            <TabController.TabPage key={key} lazy={index !== 0} index={index++}>
+              <View paddingT-20 flex style={pageStyle}>
+                <FlatList
+                  showsVerticalScrollIndicator={false}
+                  data={section.screens}
+                  keyExtractor={item => (item.screen ? item.title : `header_${item.title}`)}
+                  renderItem={this.renderItem}
+                />
+              </View>
+            </TabController.TabPage>
+          );
+        })}
+      </TabController.PageCarousel>
     );
   }
 
@@ -368,48 +236,33 @@ export default class MainScreen extends Component {
     const data = this.getMenuData();
 
     return (
-      <View testID="demo_main_screen" flex bg-dark80 style={containerStyle}>
-        {this.renderHeader()}
+      <View testID="demo_main_screen" flex style={containerStyle}>
+        {this.renderSearch()}
+
+        {showResults && this.renderSearchResults(filteredNavigationData)}
+
+        {showCarousel && (
+          <TabController asCarousel>
+            <TabController.TabBar items={_.map(data, section => ({label: section.title}))}/>
+            {this.renderPages(data)}
+          </TabController>
+        )}
         {showNoResults && (
-          <View paddingH-24>
+          <View padding-20>
             <Text dark40 text50>
               Sorry, nothing was found. Try Button or something..
             </Text>
           </View>
         )}
-        {showCarousel && (
-          <View flex useSafeArea>
-            {this.renderBreadcrumbs()}
-            {this.renderCarousel(data)}
-          </View>
-        )}
-        {showResults && this.renderSearchResults(filteredNavigationData)}
       </View>
     );
   }
 }
 
 const styles = StyleSheet.create({
-  breadcrumbs: {
-    padding: 12,
-    flexWrap: 'wrap',
-    flexDirection: 'row',
-    justifyContent: 'center'
-  },
-  page: {
-    flex: 1,
-    paddingLeft: 24
-  },
-  pageTitleContainer: {
-    borderBottomWidth: 1,
-    paddingBottom: 4,
-    borderColor: Colors.dark60
-  },
-  pageTitleExtraDivider: {
-    // marginBottom: 22,
-    marginTop: 5
-  },
   entryTextDeprecated: {
     textDecorationLine: 'line-through'
   }
 });
+
+export default gestureHandlerRootHOC(MainScreen);
