@@ -1,7 +1,6 @@
 import React, {Component} from 'react';
 import {StyleSheet, ScrollView} from 'react-native';
-import {View, Text, TextField, Colors, Spacings} from 'react-native-ui-lib';
-import {Navigation} from 'react-native-navigation';
+import {View, TextField, Colors, Spacings} from 'react-native-ui-lib';
 
 class CustomInputsScreen extends Component {
   state = {};
@@ -9,9 +8,6 @@ class CustomInputsScreen extends Component {
     return (
       <ScrollView>
         <View padding-s5>
-          <Text text40 marginB-s5>
-            Custom Inputs
-          </Text>
           <TextField title="Default" placeholder="Enter Text" enableErrors={false} containerStyle={styles.input}/>
           <TextField
             title="Square"
@@ -43,12 +39,10 @@ class CustomInputsScreen extends Component {
   }
 }
 
-export default CustomInputsScreen;
-
-Navigation.registerComponent('unicorn.components.CustomInputsScreen', () => CustomInputsScreen);
-
 const styles = StyleSheet.create({
   input: {
     marginBottom: Spacings.s4
   }
 });
+
+export default CustomInputsScreen;
