@@ -1,45 +1,60 @@
-import {Navigation} from 'react-native-navigation';
 
-Navigation.registerComponent('unicorn.components.ActionSheetScreen', () => require('./ActionSheetScreen').default);
-Navigation.registerComponent('unicorn.components.ActionBarScreen', () => require('./ActionBarScreen').default);
-Navigation.registerComponent('unicorn.components.AvatarsScreen', () => require('./AvatarsScreen').default);
-Navigation.registerComponent('unicorn.components.AnimatedImageScreen', () => require('./AnimatedImageScreen').default);
-Navigation.registerComponent('unicorn.components.ButtonsScreen', () => require('./ButtonsScreen').default);
-Navigation.registerComponent('unicorn.components.BadgesScreen', () => require('./BadgesScreen').default);
-Navigation.registerComponent('unicorn.components.CardsScreen', () => require('./CardsScreen').default);
-Navigation.registerComponent('unicorn.components.CarouselScreen', () => require('./CarouselScreen').default);
-Navigation.registerComponent('unicorn.components.CheckboxScreen', () => require('./CheckboxScreen').default);
-Navigation.registerComponent('unicorn.components.ConnectionStatusBar', () => require('./ConnectionStatusBarScreen').default);
-Navigation.registerComponent('unicorn.components.DialogScreen', () => require('./DialogScreen').default);
-Navigation.registerComponent('unicorn.components.DrawerScreen', () => require('./DrawerScreen').default);
-Navigation.registerComponent('unicorn.components.TagsInputScreen', () => require('./TagsInputScreen').default);
-Navigation.registerComponent('unicorn.components.HintsScreen', () => require('./HintsScreen').default);
-Navigation.registerComponent('unicorn.components.ImageScreen', () => require('./ImageScreen').default);
-Navigation.registerComponent('unicorn.components.KeyboardAwareScrollViewScreen', () => require('./KeyboardAwareScrollViewScreen').default);
-Navigation.registerComponent('unicorn.components.MaskedInputScreen', () => require('./MaskedInputScreen').default);
-Navigation.registerComponent('unicorn.components.OverlaysScreen', () => require('./OverlaysScreen').default);
-Navigation.registerComponent('unicorn.components.PageControlScreen', () => require('./PageControlScreen').default);
-Navigation.registerComponent('unicorn.components.PanDismissibleScreen', () => require('./PanDismissibleScreen').default);
-Navigation.registerComponent('unicorn.components.PanListenerScreen', () => require('./PanListenerScreen').default);
-Navigation.registerComponent('unicorn.components.PanResponderScreen', () => require('./PanResponderScreen').default);
-Navigation.registerComponent('unicorn.components.PickerScreen', () => require('./PickerScreen').default);
-Navigation.registerComponent('unicorn.components.RadioButtonScreen', () => require('./RadioButtonScreen').default);
-Navigation.registerComponent('unicorn.components.ScrollBarScreen', () => require('./ScrollBarScreen').default);
-Navigation.registerComponent('unicorn.components.SharedTransitionScreen', () => require('./SharedTransitionScreen').default);
-Navigation.registerComponent('unicorn.components.StepperScreen', () => require('./StepperScreen').default);
-Navigation.registerComponent('unicorn.components.SwitchScreen', () => require('./SwitchScreen').default);
-Navigation.registerComponent('unicorn.components.ToastsScreen', () => require('./ToastsScreen').default);
-Navigation.registerComponent('unicorn.components.TabBarScreen', () => require('./TabBarScreen').default);
-Navigation.registerComponent('unicorn.components.TextScreen', () => require('./TextScreen').default);
-Navigation.registerComponent('unicorn.components.TextFieldScreen', () => require('./TextFieldScreen').default);
-Navigation.registerComponent('unicorn.components.TourScreen', () => require('./TourScreen').default);
-Navigation.registerComponent('unicorn.components.FeatureHighlightScreen', () => require('./FeatureHighlightScreen').default);
-Navigation.registerComponent('unicorn.components.WheelPickerDialogScreen', () => require('./WheelPickerDialogScreen').default);
-Navigation.registerComponent('unicorn.components.SliderScreen', () => require('./SliderScreen').default);
-Navigation.registerComponent('unicorn.components.FloatingButtonScreen', () => require('./FloatingButtonScreen').default);
-Navigation.registerComponent('unicorn.components.ColorPickerScreen', () => require('./ColorPickerScreen').default);
-Navigation.registerComponent('unicorn.components.ColorSwatchScreen', () => require('./ColorSwatchScreen').default);
-Navigation.registerComponent('unicorn.components.StackAggregatorScreen', () => require('./StackAggregatorScreen').default);
-Navigation.registerComponent('unicorn.components.DateTimePickerScreen', () => require('./DateTimePickerScreen').default);
-Navigation.registerComponent('unicorn.components.ViewScreen', () => require('./ViewScreen').default);
-Navigation.registerComponent('unicorn.components.WizardScreen', () => require('./WizardScreen').default);
+export function registerScreens(registrar) {
+
+  registrar('unicorn.components.ActionSheetScreen', () => require('./ActionSheetScreen').default);
+  registrar('unicorn.components.ActionBarScreen', () => require('./ActionBarScreen').default);
+  registrar('unicorn.components.AvatarsScreen', () => require('./AvatarsScreen').default);
+  registrar('unicorn.components.AnimatedImageScreen', () => require('./AnimatedImageScreen').default);
+  registrar('unicorn.components.ButtonsScreen', () => require('./ButtonsScreen').default);
+  registrar('unicorn.components.BadgesScreen', () => require('./BadgesScreen').default);
+  registrar('unicorn.components.CardsScreen', () => require('./CardsScreen').default);
+  registrar('unicorn.animations.CardScannerScreen', () => require('../componentScreens/CardScannerScreen').default);
+  registrar('unicorn.components.CarouselScreen', () => require('./CarouselScreen').default);
+  registrar('unicorn.components.CheckboxScreen', () => require('./CheckboxScreen').default);
+  registrar('unicorn.components.ConnectionStatusBar', () => require('./ConnectionStatusBarScreen').default);
+  registrar('unicorn.components.DialogScreen', () => require('./DialogScreen').default);
+  registrar('unicorn.components.DrawerScreen', () => require('./DrawerScreen').default);
+  registrar('unicorn.components.TagsInputScreen', () => require('./TagsInputScreen').default);
+  registrar('unicorn.components.HintsScreen', () => require('./HintsScreen').default);
+  registrar('unicorn.components.ImageScreen', () => require('./ImageScreen').default);
+  registrar('unicorn.components.KeyboardAwareScrollViewScreen', () => require('./KeyboardAwareScrollViewScreen').default);
+  registrar('unicorn.components.MaskedInputScreen', () => require('./MaskedInputScreen').default);
+  registrar('unicorn.components.OverlaysScreen', () => require('./OverlaysScreen').default);
+  registrar('unicorn.components.PageControlScreen', () => require('./PageControlScreen').default);
+  registrar('unicorn.components.PanDismissibleScreen', () => require('./PanDismissibleScreen').default);
+  registrar('unicorn.components.PanListenerScreen', () => require('./PanListenerScreen').default);
+  registrar('unicorn.components.PanResponderScreen', () => require('./PanResponderScreen').default);
+  registrar('unicorn.components.PickerScreen', () => require('./PickerScreen').default);
+  registrar('unicorn.animations.ProgressBarScreen', () => require('../componentScreens/ProgressBarScreen').default);
+  registrar('unicorn.components.RadioButtonScreen', () => require('./RadioButtonScreen').default);
+  registrar('unicorn.components.ScrollBarScreen', () => require('./ScrollBarScreen').default);
+  registrar('unicorn.components.SharedTransitionScreen', () => require('./SharedTransitionScreen').default);
+  registrar('unicorn.components.StepperScreen', () => require('./StepperScreen').default);
+  registrar('unicorn.components.SwitchScreen', () => require('./SwitchScreen').default);
+  registrar('unicorn.components.ToastsScreen', () => require('./ToastsScreen').default);
+  registrar('unicorn.components.TabControllerScreen', () => require('./TabControllerScreen').default);
+  registrar('unicorn.components.TabBarScreen', () => require('./TabBarScreen').default);
+  registrar('unicorn.components.TextScreen', () => require('./TextScreen').default);
+  registrar('unicorn.components.TextFieldScreen', () => require('./TextFieldScreen').default);
+  registrar('unicorn.wrappers.TouchableOpacityScreen', () => require('./TouchableOpacityScreen').default);
+  registrar('unicorn.components.TourScreen', () => require('./TourScreen').default);
+  registrar('unicorn.components.FeatureHighlightScreen', () => require('./FeatureHighlightScreen').default);
+  registrar('unicorn.components.WheelPickerDialogScreen', () => require('./WheelPickerDialogScreen').default);
+  registrar('unicorn.components.SliderScreen', () => require('./SliderScreen').default);
+  registrar('unicorn.components.FloatingButtonScreen', () => require('./FloatingButtonScreen').default);
+  registrar('unicorn.components.ColorPickerScreen', () => require('./ColorPickerScreen').default);
+  registrar('unicorn.components.ColorSwatchScreen', () => require('./ColorSwatchScreen').default);
+  registrar('unicorn.components.StackAggregatorScreen', () => require('./StackAggregatorScreen').default);
+  registrar('unicorn.components.DateTimePickerScreen', () => require('./DateTimePickerScreen').default);
+  registrar('unicorn.components.ViewScreen', () => require('./ViewScreen').default);
+  registrar('unicorn.components.WizardScreen', () => require('./WizardScreen').default);
+  // List Components
+  registrar('unicorn.lists.BasicListScreen', () => require('./BasicListScreen').default);
+  registrar('unicorn.lists.ContactsListScreen', () => require('./ContactsListScreen').default);
+  registrar('unicorn.lists.ConversationListScreen', () => require('./ConversationListScreen').default);
+  // Full Screen components
+  registrar('unicorn.screens.EmptyStateScreen', () => require('./EmptyStateScreen').default);
+  registrar('unicorn.screens.LoadingScreen', () => require('./LoadingScreen').default);
+  registrar('unicorn.screens.ModalScreen', () => require('./ModalScreen').default);
+}
+
