@@ -1,24 +1,16 @@
-import {Navigation} from 'react-native-navigation';
 import {gestureHandlerRootHOC} from 'react-native-gesture-handler';
 
-import './componentScreens';
-import './nativeComponentScreens';
-import './componentScreenScreens';
-import './listScreens';
-import './styleScreens';
-import './animationScreens';
-import './realExamples';
-import './wrapperScreens';
-import './incubatorScreens';
+export function registerScreens(registrar) {
 
-import MainScreen from './MainScreen';
-import Playground from './PlaygroundScreen';
-import SettingsScreen from './SettingsScreen';
-import CustomScreen from './CustomScreen';
+  require('./componentScreens').registerScreens(registrar);
+  require('./foundationScreens').registerScreens(registrar);
+  require('./nativeComponentScreens').registerScreens(registrar);
+  require('./animationScreens').registerScreens(registrar);
+  require('./realExamples').registerScreens(registrar);
+  require('./incubatorScreens').registerScreens(registrar);
 
-export function registerScreens() {
-  Navigation.registerComponent('unicorn.MainScreen', () => MainScreen);
-  Navigation.registerComponent('unicorn.PlaygroundScreen', () => gestureHandlerRootHOC(Playground));
-  Navigation.registerComponent('unicorn.Settings', () => SettingsScreen);
-  Navigation.registerComponent('unicorn.CustomScreen', () => CustomScreen);
+  registrar('unicorn.MainScreen', () => require('./MainScreen').default);
+  registrar('unicorn.PlaygroundScreen', () => gestureHandlerRootHOC(require('./PlaygroundScreen').default));
+  registrar('unicorn.Settings', () => require('./SettingsScreen').default);
+  registrar('unicorn.CustomScreen', () => require('./CustomScreen').default);
 }
