@@ -6,6 +6,6 @@ export * from './helpers';
 export * from './incubator';
 export * from './interactableComponents';
 export * from './nativeComponents';
-export * from './screensComponents';
 export * from './services';
-export * from './style';
+export * from './generatedTypes';
+export * from './globalTypes';
